+++ R/data.r
@@ -2,11 +2,12 @@
 #'
 #' @description
 #' The table *batch_keywords* contains the keywords for each batch. Each
-#' line contains one *keyword*, the *type* of the batch (i.e. control
+#' line contains one *keyword*, the *type* of the batch (i.e., control
 #' or object) and the id of the *batch* to which the keyword is assigned.
 #' Keywords can be added with the function `add_keywords`. The function
 #' `start_db` exports the table *batch_keywords* as objects
-#' `keywords_control` and `keywords_object` to `.GlobalEnv`.
+#' `keywords_control` and `keywords_object` to the package environment
+#' `gt.env`.
 #'
 #' Example data for the table *batch_keywords* is available as R object
 #' `example_keywords`.
@@ -31,7 +32,7 @@
 #' @description
 #' The table *batch_time* contains the time period for which data is
 #' downloaded for each batch. Each line contains one *time* period, the
-#' *type* of the batch (i.e. control or object) and the id of the
+#' *type* of the batch (i.e., control or object) and the id of the
 #' *batch* to which the time period is assigned. Time frames take the form
 #' `"YYYY-MM-DD YYYY-MM-DD"`. Time periods are added automatically through
 #' the function `add_keywords`. The function `start_db` exports the
@@ -64,8 +65,8 @@
 #' gets the value *world* as location. Data is downloaded and automatically
 #' written to the table through the function `download_control`. The
 #' function `start_db` exports the table *data_control* as database
-#' connection `tbl_control` to `.GlobalEnv`. Users can access the
-#' database table through `dplyr::tbl`.
+#' connection `tbl_control` to the package environment `gt.env`. Users
+#' can access the database table through `dplyr::tbl`.
 #' The sample data included in `data_control` was simulated based on actual
 #' Google Trends data.
 #'
@@ -103,8 +104,8 @@
 #' mapped. Global data takes the value *world* as location. Data is
 #' downloaded and automatically written to the table through the function
 #' `download_object`. The function `start_db` exports the table
-#' *data_object* as database connection `tbl_object` to
-#' `.GlobalEnv`. Users can access the database table through
+#' *data_object* as database connection `tbl_object` to the package
+#' environment `gt.env`. Users can access the database table through
 #' `dplyr::tbl`.
 #' The sample data included in `data_object` was simulated based on actual
 #' Google Trends data.
@@ -148,8 +149,8 @@
 #' scores are computed and automatically written to the table with the function
 #' `compute_score`. The function `start_db` exports the table
 #' *data_score* as database connection `tbl_score` to
-#' `.GlobalEnv`. Users can access the database table through
-#' `dplyr::tbl`.
+#' the package environment `gt.env`. Users can access the database
+#' table through `dplyr::tbl`.
 #' The sample data included in `data_score` was simulated based on actual
 #' Google Trends data.
 #'
@@ -199,8 +200,8 @@
 #' distribution of search scores. DOI is computed and automatically written to
 #' the table with the function `compute_doi`. The function `start_db`
 #' exports the table *data_doi* as database connection `tbl_doi` to
-#' `.GlobalEnv`. Users can access the database table through
-#' `dplyr::tbl`.
+#' the package environment `gt.env`. Users can access the database table
+#' through `dplyr::tbl`.
 #' The sample data included in `data_doi` was simulated based on actual
 #' Google Trends data.
 #'