+++ tests/testthat/test-downloads.R
@@ -8,19 +8,6 @@ Sys.setenv("LANGUAGE" = "EN")
 initialize_db()
 start_db()
 
-rm(
-  tbl_control,
-  tbl_doi,
-  tbl_object,
-  tbl_score,
-  keyword_synonyms,
-  keywords_control,
-  keywords_object,
-  time_control,
-  time_object,
-  envir = .GlobalEnv
-)
-
 add_control_keyword(
   keyword = c("gmail", "map", "translate", "wikipedia", "youtube"),
   time = "2010-01-01 2019-12-31"
@@ -34,14 +21,17 @@ add_object_keyword(
 # try download object ----------------------------------------------------------
 test_that("download_object1", {
   out <- expect_message(
-    download_object(object = 1, locations = countries[[1]]),
+    download_object(object = 1, locations = gt.env$countries[[1]]),
     "Download for object data failed.\nThere is no data in 'data_control' for control batch 1 and location US."
   )
 })
 
 # download control --------------------------------------------------------------
 test_that("download_control1", {
-  out <- capture_messages(download_control(control = 1, locations = countries[1:3]))
+  skip_on_cran()
+  skip_if_offline()
+  
+  out <- capture_messages(download_control(control = 1, locations = gt.env$countries[1:3]))
 
   expect_match(
     out,
@@ -59,26 +49,32 @@ test_that("download_control1", {
     all = FALSE
   )
 
-  out <- filter(.tbl_control, batch == 1 & location != "world")
+  out <- filter(gt.env$tbl_control, batch == 1 & location != "world")
   out <- collect(out)
   expect_equal(nrow(out), 1800)
 })
 
 # re-download control ----------------------------------------------------------
 test_that("download_control2", {
+  skip_on_cran()
+  skip_if_offline()
+  
   expect_message(
-    download_control(control = 1, locations = countries[[1]]),
+    download_control(control = 1, locations = gt.env$countries[[1]]),
     "Control data already available | control: 1 | location: US [1/1]"
   )
 })
 
 # download control global ------------------------------------------------------
 test_that("download_control3", {
+  skip_on_cran()
+  skip_if_offline()
+  
   expect_message(
     download_control_global(control = 1),
     "Successfully downloaded control data | control: 1 | location: world [1/1]"
   )
-  out <- filter(.tbl_control, batch == 1 & location == "world")
+  out <- filter(gt.env$tbl_control, batch == 1 & location == "world")
   out <- collect(out)
   expect_equal(nrow(out), 600)
 })
@@ -109,7 +105,10 @@ test_that("download_control6", {
 
 # download object --------------------------------------------------------------
 test_that("download_object2", {
-  out <- capture_messages(download_object(object = 1, locations = countries[1:3]))
+  skip_on_cran()
+  skip_if_offline()
+  
+  out <- capture_messages(download_object(object = 1, locations = gt.env$countries[1:3]))
 
   expect_match(
     out,
@@ -127,26 +126,32 @@ test_that("download_object2", {
     all = FALSE
   )
 
-  out <- filter(.tbl_object, batch_o == 1 & location != "world")
+  out <- filter(gt.env$tbl_object, batch_o == 1 & location != "world")
   out <- collect(out)
   expect_equal(nrow(out), 1800)
 })
 
 # re-download object -----------------------------------------------------------
 test_that("download_object3", {
+  skip_on_cran()
+  skip_if_offline()
+  
   expect_message(
-    download_object(object = 1, locations = countries[[1]]),
+    download_object(object = 1, locations = gt.env$countries[[1]]),
     "Object data already available | object: 1 | control: 1 | location: US [1/1]"
   )
 })
 
 # download object global -------------------------------------------------------
 test_that("download_object4", {
+  skip_on_cran()
+  skip_if_offline()
+  
   expect_message(
     download_object_global(object = 1),
     "Successfully downloaded object data | object: 1 | control: 1 | location: world [1/1]"
   )
-  out <- filter(.tbl_object, batch_o == 1 & location == "world")
+  out <- filter(gt.env$tbl_object, batch_o == 1 & location == "world")
   out <- collect(out)
   expect_equal(nrow(out), 600)
 })