+++ tests/testthat/test-locations.R
@@ -19,9 +19,9 @@ test_that("add_loc1", {
     "Successfully created new location set dach \\(AT, DE, CH\\)\\."
   )
 
-  expect_error(
-    dach,
-    "object 'dach' not found"
+  expect_equal(
+    gt.env$dach,
+    NULL
   )
 
   expect_message(
@@ -34,7 +34,7 @@ test_that("add_loc1", {
   )
 
   expect_identical(
-    asia,
+    gt.env$asia,
     c("CN", "JP")
   )
 })
@@ -44,12 +44,12 @@ start_db()
 
 test_that("add_loc2", {
   expect_identical(
-    dach,
+    gt.env$dach,
     c("AT", "DE", "CH")
   )
 
   expect_identical(
-    asia,
+    gt.env$asia,
     c("CN", "JP")
   )
 })
@@ -65,12 +65,12 @@ add_object_keyword(
   time = "2010-01-01 2019-12-31"
 )
 
-dbWriteTable(globaltrends_db, "data_control", example_control, append = TRUE)
-dbWriteTable(globaltrends_db, "data_object", example_object, append = TRUE)
+dbWriteTable(gt.env$globaltrends_db, "data_control", example_control, append = TRUE)
+dbWriteTable(gt.env$globaltrends_db, "data_object", example_object, append = TRUE)
 
 # compute score ----------------------------------------------------------------
 test_that("compute_score1", {
-  out <- capture_messages(compute_score(object = 1, control = 1, locations = asia))
+  out <- capture_messages(compute_score(object = 1, control = 1, locations = gt.env$asia))
 
   expect_match(
     out,
@@ -83,7 +83,7 @@ test_that("compute_score1", {
     all = FALSE
   )
 
-  out <- tbl_score %>%
+  out <- gt.env$tbl_score %>%
     count(location) %>%
     collect()
 
@@ -99,9 +99,9 @@ test_that("compute_score1", {
 })
 
 test_that("compute_score2", {
-  compute_score(object = 1, control = 1, locations = countries)
+  compute_score(object = 1, control = 1, locations = gt.env$countries)
 
-  out <- tbl_score %>%
+  out <- gt.env$tbl_score %>%
     count(location) %>%
     collect()
 
@@ -125,7 +125,7 @@ test_that("compute_doi", {
     "Successfully computed DOI | control: 1 | object: 1 [1/1]"
   )
 
-  out <- tbl_doi %>%
+  out <- gt.env$tbl_doi %>%
     count(locations) %>%
     collect()
 
@@ -139,7 +139,7 @@ test_that("compute_doi", {
     c(1440, 1440)
   )
 
-  out <- tbl_doi %>%
+  out <- gt.env$tbl_doi %>%
     group_by(locations) %>%
     summarise(
       gini = mean(gini, na.rm = TRUE),
@@ -155,8 +155,8 @@ test_that("compute_doi", {
 
 # export score -----------------------------------------------------------------
 test_that("export_score", {
-  out1 <- export_score(location = asia)
-  out2 <- export_score(location = countries)
+  out1 <- export_score(location = gt.env$asia)
+  out2 <- export_score(location = gt.env$countries)
 
   expect_equal(
     unique(out1$location),
@@ -196,13 +196,85 @@ test_that("plot_box", {
   expect_false(identical(plot1, plot2))
 })
 
+# namibia ----------------------------------------------------------------------
+test_that("namibia1", {
+  expect_message(
+    add_locations("NA", "test"),
+    "Successfully created new location set test \\(NA\\)\\."
+  )
+  
+  out <- gt.env$tbl_locations %>%
+    filter(type == "test") %>%
+    collect() %>%
+    pull(location)
+  expect_equal(out, "NX")
+})
+
+test_that("namibia2", {
+  skip_on_cran()
+  skip_if_offline()
+  
+  add_control_keyword(keyword = "google", time = "2010-01-01 2020-12-31")
+  expect_message(
+    download_control(control = 2, locations = gt.env$test),
+    "Successfully downloaded control data | control: 2 | location: NA [1/1]"
+  )
+  
+  out <- gt.env$tbl_control %>%
+    filter(batch == 2) %>%
+    collect()
+  expect_equal(nrow(out), 132)
+  expect_equal(out$location[[1]], "NX")
+  
+  out <- export_control(location = gt.env$test, control = 2)
+  expect_equal(nrow(out), 132)
+  expect_equal(out$location[[1]], "NA")
+})
+
+test_that("namibia3", {
+  skip_on_cran()
+  skip_if_offline()
+  
+  add_object_keyword(keyword = "football", time = "2010-01-01 2020-12-31")
+  expect_message(
+    download_object(object = 2, control = 2, locations = gt.env$test),
+    "Successfully downloaded object data | object: 2 | control: 2 | location: NA [1/1]"
+  )
+  
+  out <- gt.env$tbl_object %>%
+    filter(batch_c == 2) %>%
+    collect()
+  expect_equal(nrow(out), 264)
+  expect_equal(out$location[[1]], "NX")
+  
+  out <- export_object(location = gt.env$test, control = 2)
+  expect_equal(nrow(out), 264)
+  expect_equal(out$location[[1]], "NA") 
+})
+
+test_that("namibia4", {
+  skip_on_cran()
+  skip_if_offline()
+  
+  expect_message(
+    compute_score(object = 2, control = 2, locations = gt.env$test),
+    "Successfully computed search score | control: 2 | object: 2 | location: NA [1/1]"
+  )
+  
+  out <- gt.env$tbl_score %>%
+    filter(batch_c == 2) %>%
+    collect()
+  expect_equal(nrow(out), 132)
+  expect_equal(out$location[[1]], "NX")
+  
+  out <- export_score(location = gt.env$test, control = 2)
+  expect_equal(nrow(out), 132)
+  expect_equal(out$location[[1]], "NA") 
+})
+
 # signals ----------------------------------------------------------------------
 test_that("signals1", {
   test_locations(fun = add_locations, type = "A")
-})
-
-test_that("signals2", {
-  test_type(fun = add_locations, incl = c(1, 3:5), locations = "A")
 })
 
 test_that("signals3", {