+++ NEWS.md
@@ -1,3 +1,8 @@
+# globaltrends v.0.0.12
+* Stop direct exports from functions to .GlobalEnv
+* Exports of objects (e.g., SQL connections, keyword tables) are redirected to package environment `gt.env`
+* Add handle for Namibia ISO code
+
 # globaltrends v.0.0.11
 * Add function plot_map()
 * Add automatic stringr::str_squish for keywords and synonyms